requirements.txt
get==2019.4.13
post==2019.4.13
public==2019.4.13
PyMySQL==0.9.3
query-string==2019.4.13
request==2019.4.13
schedule==0.6.0
selenium==3.141.0
urllib3==1.25.1
